feat: move bookkeeping to Tvb & TreeItem
Tvb struct is cheap to copy around, this should be where TreeItem
should represent the iteam we add to the protocol tree, with its
own 'view' of the buffer behind it.

Tree object itself should be cointaining the main top level view of the
buffer. i.e. it should move its offset based on Items and subtrees you
add.