fix: wireshark crash on empty display string
When you hit enter on the display filter on the empty string,
Wireshark validates packets against all dissectors, but it calls it
with a NULL proto_tree arguement. This part is documented under  the
optimisation section of the README.dissector which I missed handling.

In this case, I think it reasonable to avoid building a tree in such a
case in an opinionated frame work.

An excerpt of the wireshark docs below:
Note also that there is no guarantee, the first time the dissector is
called, whether "tree" will be null or not; your dissector must work
correctly, building or updating whatever state information is
necessary, in either case.